requirements.txt
google-adk
chromadb
pypdf2
python-dotenv
sentence-transformers